Extract MultiParentCasperMergeSpec
